+++ src/proposals/MIPProposal.s.sol
@@ -63,7 +63,7 @@ abstract contract MIPProposal is Script {
         addresses = new Addresses();
         vm.makePersistent(address(addresses));
 
-        vm.selectFork(primaryForkId());
+        vm.selectFork(uint256(primaryForkId()));
 
         address deployerAddress = vm.addr(PRIVATE_KEY);
